add get_latest_utils to transform utils for unit testing.
